+++ config/Router.php
@@ -3,6 +3,7 @@
 namespace BrunoGrosdidier\Blog\config;
 
 use BrunoGrosdidier\Blog\src\controller\FrontController;
+use BrunoGrosdidier\Blog\src\controller\BackController;
 use BrunoGrosdidier\Blog\src\controller\ErrorController;
 use Exception;
 
@@ -10,18 +11,18 @@ Class Router
 {
 
 	private $frontController;
+    private $backController;
 	private $errorController;
 
     public function __construct()
     {
         $this->frontController = new FrontController();
+        $this->backController = new BackController();
         $this->errorController = new ErrorController();
     }
 
 	public function run()
 	{
-
-
 	    try {
 			if (isset($_GET['action'])) {
 				if ($_GET['action'] == 'getAllPosts') {
@@ -38,7 +39,7 @@ Class Router
 				elseif ($_GET['action'] == 'addOneComment') {
 					if (isset($_GET['id']) && $_GET['id'] > 0) {
 						if (!empty($_POST['author']) && !empty($_POST['comment'])) {
-							$this->frontController->addOneComment($_GET['id'], $_POST['author'], $_POST['comment']);							
+							$this->backController->addOneComment($_GET['id'], $_POST['author'], $_POST['comment']);
 						}
 						else {
 							throw new Exception('Tous les champs ne sont pas remplis !');
@@ -50,7 +51,7 @@ Class Router
 				}
 				elseif ($_GET['action'] == 'editOneComment') {
 					if (isset($_GET['commentId']) && $_GET['commentId'] > 0) {
-						$this->frontController->editOneComment();						
+						$this->backController->editOneComment();
 					}
 					else {
 						throw new Exception('Aucun identifiant de commentaire envoyé');
@@ -60,7 +61,7 @@ Class Router
 					if (isset($_GET['postId']) && $_GET['postId'] > 0) {
 						if (isset($_GET['commentId']) && $_GET['commentId'] > 0) {
 							if (!empty($_POST['commentText'])) {
-								$this->frontController->refreshOneComment($_GET['commentId'], $_POST['commentText'], $_GET['postId']);								
+								$this->backController->refreshOneComment($_GET['commentId'], $_POST['commentText'], $_GET['postId']);
 							}
 							else {
 								throw new Exception('Le champ n\'est pas rempli !');
@@ -71,12 +72,9 @@ Class Router
 						}
 					}
 					else {
-						throw new Exception('Putain ! Aucun identifiant de billet envoyé');
+						throw new Exception('Aucun identifiant de billet envoyé');
 					}			
 				}
-                else {
-                    var_dump('coucou');
-                }
 			}
 			else {
 				$this->frontController->getAllPosts();				
@@ -86,31 +84,5 @@ Class Router
 			echo 'Erreur : ' . $e->getMessage();
             $this->errorController->errorServer();
 		}
-
-
-        //var_dump($_GET);
-        /*
-        try{
-            if(isset($_GET['action']))
-            {
-                //if($_GET['action'] === 'getAllPosts'){
-                    //$this->frontController->getOnePost($_GET['postId']);
-                //}
-                //elseif($_GET['action'] === 'addPost'){
-                    //$this->frontController->addArticle($_POST);
-                //}
-                //else{
-                    //$this->errorController->errorNotFound();
-                //}
-            }
-            else{
-                $this->frontController->getAllPosts();
-            }
-        }
-        catch (Exception $e)
-        {
-            $this->errorController->errorServer();
-        }
-        */
 	}
 }
